\name{oral.Ds}
\alias{oral.Ds}
\docType{data}
\title{
A list of ecological (pairwise, subject-by-subject, n by n) distance matrices for the oral microbiome data (Park et al., 2023, BMC Microbiology)
}
\description{
A list of ecological (pairwise, subject-by-subject, n by n) distance matrices for the oral microbiome data (Park et al., 2023, BMC Microbiology)
}
\usage{data("oral.Ds")}
\format{
  The format is:
List of 7
 $ Jaccard       : num [1:57, 1:57] 0 0.563 0.566 0.574 0.551 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
 $ Bray-Curtis   : num [1:57, 1:57] 0 0.615 0.653 0.705 0.653 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
 $ UUniFrac      : num [1:57, 1:57] 0 0.149 0.2 0.191 0.187 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
 $ GUniFrac(0.25): num [1:57, 1:57] 0 0.353 0.341 0.399 0.422 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
 $ GUniFrac(0.5) : num [1:57, 1:57] 0 0.274 0.246 0.313 0.334 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
 $ GUniFrac(0.75): num [1:57, 1:57] 0 0.208 0.18 0.232 0.25 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
 $ WUniFrac      : num [1:57, 1:57] 0 0.151 0.128 0.158 0.174 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
  .. ..$ : chr [1:57] "V108" "V109" "V112" "V113" ...
}
\references{
Park B, Koh H, Patatanian M, Reyes-Caballero H, Zhao N, Meinert J, Holbrook JT, Leinbach LI, Biswal S. The mediating roles of the oral microbiome in saliva and subgingival sites between e-cigarette smoking and gingival inflammation. BMC Microbiol, 23(35), 2023

Koh H. An ensemble learning method for joint kernel association testing and principal component analysis on multiple kernels. (In Review)
}

\examples{
data(oral.Ds)
}
\keyword{datasets}
