\name{oral.otu.tab}
\alias{oral.otu.tab}
\docType{data}
\title{
A feature table for the oral microbiome data (Park et al., 2023, BMC Microbiology)
}
\description{
A feature table for the oral microbiome data (Park et al., 2023, BMC Microbiology)
}
\usage{data("oral.otu.tab")}
\format{
  A data frame with 2266 observations on the following 57 variables.
  \describe{
    \item{\code{V108}}{a numeric vector}
    \item{\code{V109}}{a numeric vector}
    \item{\code{V112}}{a numeric vector}
    \item{\code{V113}}{a numeric vector}
    \item{\code{V116}}{a numeric vector}
    \item{\code{V117}}{a numeric vector}
    \item{\code{V118}}{a numeric vector}
    \item{\code{V120}}{a numeric vector}
    \item{\code{V126}}{a numeric vector}
    \item{\code{V130}}{a numeric vector}
    \item{\code{V134}}{a numeric vector}
    \item{\code{V135}}{a numeric vector}
    \item{\code{V150}}{a numeric vector}
    \item{\code{V169}}{a numeric vector}
    \item{\code{V181}}{a numeric vector}
    \item{\code{V184}}{a numeric vector}
    \item{\code{V186}}{a numeric vector}
    \item{\code{V188}}{a numeric vector}
    \item{\code{V191}}{a numeric vector}
    \item{\code{V193}}{a numeric vector}
    \item{\code{V194}}{a numeric vector}
    \item{\code{V196}}{a numeric vector}
    \item{\code{V201}}{a numeric vector}
    \item{\code{V203}}{a numeric vector}
    \item{\code{V204}}{a numeric vector}
    \item{\code{V217}}{a numeric vector}
    \item{\code{V233}}{a numeric vector}
    \item{\code{V237}}{a numeric vector}
    \item{\code{V241}}{a numeric vector}
    \item{\code{V248}}{a numeric vector}
    \item{\code{V251}}{a numeric vector}
    \item{\code{V258}}{a numeric vector}
    \item{\code{V262}}{a numeric vector}
    \item{\code{V265}}{a numeric vector}
    \item{\code{V276}}{a numeric vector}
    \item{\code{V282}}{a numeric vector}
    \item{\code{V285}}{a numeric vector}
    \item{\code{V306}}{a numeric vector}
    \item{\code{V311}}{a numeric vector}
    \item{\code{V314}}{a numeric vector}
    \item{\code{V319}}{a numeric vector}
    \item{\code{V326}}{a numeric vector}
    \item{\code{V327}}{a numeric vector}
    \item{\code{V328}}{a numeric vector}
    \item{\code{V330}}{a numeric vector}
    \item{\code{V331}}{a numeric vector}
    \item{\code{V334}}{a numeric vector}
    \item{\code{V339}}{a numeric vector}
    \item{\code{V342}}{a numeric vector}
    \item{\code{V343}}{a numeric vector}
    \item{\code{V344}}{a numeric vector}
    \item{\code{V346}}{a numeric vector}
    \item{\code{V356}}{a numeric vector}
    \item{\code{V363}}{a numeric vector}
    \item{\code{V368}}{a numeric vector}
    \item{\code{V372}}{a numeric vector}
    \item{\code{V375}}{a numeric vector}
  }
}

\references{
Park B, Koh H, Patatanian M, Reyes-Caballero H, Zhao N, Meinert J, Holbrook JT, Leinbach LI, Biswal S. The mediating roles of the oral microbiome in saliva and subgingival sites between e-cigarette smoking and gingival inflammation. BMC Microbiol, 23(35), 2023

Koh H. An ensemble learning method for joint kernel association testing and principal component analysis on multiple kernels. (In Review)
}

\examples{
data(oral.otu.tab)
}
\keyword{datasets}
