\name{throat.otu.tab}
\alias{throat.otu.tab}
\docType{data}
\title{
A feature table for the upper-respiratory-tract microbiome data (Charlson et al., 2010, PLOS One)
}
\description{
A feature table for the upper-respiratory-tract microbiome data (Charlson et al., 2010, PLOS One)
}
\usage{data("throat.otu.tab")}
\format{
  A data frame with 740 observations on the following 57 variables.
  \describe{
    \item{\code{ESC_1.1_OPL}}{a numeric vector}
    \item{\code{ESC_1.3_OPL}}{a numeric vector}
    \item{\code{ESC_1.4_OPL}}{a numeric vector}
    \item{\code{ESC_1.5_OPL}}{a numeric vector}
    \item{\code{ESC_1.6_OPL}}{a numeric vector}
    \item{\code{ESC_1.10_OPL}}{a numeric vector}
    \item{\code{ESC_1.11_OPL}}{a numeric vector}
    \item{\code{ESC_1.12_OPL}}{a numeric vector}
    \item{\code{ESC_1.13_OPL}}{a numeric vector}
    \item{\code{ESC_1.14_OPL}}{a numeric vector}
    \item{\code{ESC_1.15_OPL}}{a numeric vector}
    \item{\code{ESC_1.18_OPL}}{a numeric vector}
    \item{\code{ESC_1.19_OPL}}{a numeric vector}
    \item{\code{ESC_1.20_OPL}}{a numeric vector}
    \item{\code{ESC_1.21_OPL}}{a numeric vector}
    \item{\code{ESC_1.23_OPL}}{a numeric vector}
    \item{\code{ESC_1.24_OPL}}{a numeric vector}
    \item{\code{ESC_1.25_OPL}}{a numeric vector}
    \item{\code{ESC_1.26_OPL}}{a numeric vector}
    \item{\code{ESC_1.27_OPL}}{a numeric vector}
    \item{\code{ESC_1.28_OPL}}{a numeric vector}
    \item{\code{ESC_1.29_OPL}}{a numeric vector}
    \item{\code{ESC_1.30_OPL}}{a numeric vector}
    \item{\code{ESC_1.31_OPL}}{a numeric vector}
    \item{\code{ESC_1.32_OPL}}{a numeric vector}
    \item{\code{ESC_1.33_OPL}}{a numeric vector}
    \item{\code{ESC_1.34_OPL}}{a numeric vector}
    \item{\code{ESC_1.35_OPL}}{a numeric vector}
    \item{\code{ESC_1.37_OPL}}{a numeric vector}
    \item{\code{ESC_1.39_OPL}}{a numeric vector}
    \item{\code{ESC_1.40_OPL}}{a numeric vector}
    \item{\code{ESC_1.42_OPL}}{a numeric vector}
    \item{\code{ESC_1.43_OPL}}{a numeric vector}
    \item{\code{ESC_1.44_OPL}}{a numeric vector}
    \item{\code{ESC_1.45_OPL}}{a numeric vector}
    \item{\code{ESC_1.46_OPL}}{a numeric vector}
    \item{\code{ESC_1.47_OPL}}{a numeric vector}
    \item{\code{ESC_1.49_OPL}}{a numeric vector}
    \item{\code{ESC_1.50_OPL}}{a numeric vector}
    \item{\code{ESC_1.51_OPL}}{a numeric vector}
    \item{\code{ESC_1.52_OPL}}{a numeric vector}
    \item{\code{ESC_1.53_OPL}}{a numeric vector}
    \item{\code{ESC_1.55_OPL}}{a numeric vector}
    \item{\code{ESC_1.56_OPL}}{a numeric vector}
    \item{\code{ESC_1.57_OPL}}{a numeric vector}
    \item{\code{ESC_1.58_OPL}}{a numeric vector}
    \item{\code{ESC_1.59_OPL}}{a numeric vector}
    \item{\code{ESC_1.60_OPL}}{a numeric vector}
    \item{\code{ESC_1.61_OPL}}{a numeric vector}
    \item{\code{ESC_1.62_OPL}}{a numeric vector}
    \item{\code{ESC_1.63_OPL}}{a numeric vector}
    \item{\code{ESC_1.64_OPL}}{a numeric vector}
    \item{\code{ESC_1.65_OPL}}{a numeric vector}
    \item{\code{ESC_1.67_OPL}}{a numeric vector}
    \item{\code{ESC_1.68_OPL}}{a numeric vector}
    \item{\code{ESC_1.69_OPL}}{a numeric vector}
    \item{\code{ESC_1.70_OPL}}{a numeric vector}
  }
}

\references{
Charlson ES, Chen J, Custers-Allen R, Bittinger K, Li H, Sinha R, Hwang J, Bushman FD, Collman RG. Disordered microbial communities in the upper respiratory tract of cigarette smokers. PLoS One, 5(12):e15216, 2010

Koh H. An ensemble learning method for joint kernel association testing and principal component analysis on multiple kernels. (In Review)
}

\examples{
data(throat.otu.tab)
}
\keyword{datasets}
